refactor(challenge): challengeColors를 도메인 공용 상수로 분리 (T1, 역결합 해소)
challengeColors가 ChallengeView(및 ChallengePortfolioView에 중복)에서 export되어
19개 파일이 페이지 컴포넌트를 색상 때문에 import하던 역결합을 제거.
- domain/program/challenge/challengeColors.ts 신설(정본, superset)
- ChallengeView / ChallengePortfolioView의 정의·중복 제거 후 import
- 색상 사용 18개 파일 import 경로를 challengeColors로 갱신
- ChallengeColor 타입도 함께 이동
tsc 0 / lint 통과. 색상값 변경 없음(ChallengeView superset 기준 통합).
