+++ apps/web/src/domain/program/challenge/challenge-view/ChallengeFaq.tsx
@@ -1,7 +1,7 @@
 import FaqDropdown from '@/common/dropdown/FaqDropdown';
 import Heading2 from '@/common/header/Heading2';
 import FaqChat from '@/domain/faq/FaqChat';
-import { challengeColors } from '@/domain/program/challenge/ChallengeView';
+import { challengeColors } from '@/domain/program/challenge/challengeColors';
 import SuperTitle from '@/domain/program/program-detail/SuperTitle';
 import { twMerge } from '@/lib/twMerge';
 import { ChallengeType, challengeTypeSchema, faqSchemaType } from '@/schema';
