+++ apps/web/src/domain/program/challenge/challenge-view/ChallengeInfoBottom.tsx
@@ -3,7 +3,7 @@ import ChevronDown from '@/assets/icons/chevron-down.svg?react';
 import ClockIcon from '@/assets/icons/clock.svg?react';
 import LaptopIcon from '@/assets/icons/laptop.svg?react';
 import Heading2 from '@/common/header/Heading2';
-import { challengeColors } from '@/domain/program/challenge/ChallengeView';
+import { challengeColors } from '@/domain/program/challenge/challengeColors';
 import BasicInfoBottomRow from '@/domain/program/program-detail/basicInfo/BasicInfoBottomRow';
 import BasicInfoRow from '@/domain/program/program-detail/basicInfo/BasicInfoRow';
 import { useInstallmentPayment } from '@/hooks/useInstallmentPayment';
