+++ apps/web/src/domain/program/challenge/challenge-view/ChallengeIntroCareerStart.tsx
@@ -1,5 +1,5 @@
 import Heading2 from '@/common/header/Heading2';
-import { challengeColors } from '@/domain/program/challenge/ChallengeView';
+import { challengeColors } from '@/domain/program/challenge/challengeColors';
 import Description from '@/domain/program/program-detail/Description';
 import { ChallengeType, challengeTypeSchema } from '@/schema';
 import { ChallengePoint } from '@/types/interface';
