+++ apps/web/src/domain/program/challenge/portfolio-view/PortfolioOneOnOne.tsx
@@ -2,7 +2,7 @@ import benefitImg1 from '@/assets/benefit1.jpg';
 import benefitImg2 from '@/assets/benefit2.jpg';
 import benefitImg3 from '@/assets/benefit3.jpg';
 import { Break } from '@/common/Break';
-import { challengeColors } from '@/domain/program/challenge/ChallengeView';
+import { challengeColors } from '@/domain/program/challenge/challengeColors';
 import SuperTitle from '@/domain/program/program-detail/SuperTitle';
 import { ChallengeType, challengeTypeSchema } from '@/schema';
 import { useMediaQuery } from '@mui/material';
